backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
openai==1.59.7
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
supabase==2.11.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
google-auth==2.38.0
email-validator==2.2.0
pypdf==6.10.2
